Update README with contract merging instructions and change damageBoostPercentage type in IBuildingManager to uint16 for consistency

+++ contracts/data/IBuildingManager.sol
@@ -9,7 +9,7 @@ interface IBuildingManager {
         uint16 prerequisiteBuildingId;
         bool isHeadquarters;
         uint16 constructionRange;
-        uint256 damageBoostPercentage; // 1-10000 (0.01% - 100%)
+        uint16 damageBoostPercentage; // 1-10000 (0.01% - 100%)
         TokenAmountLib.TokenAmount[] constructionCost;
         bool canBeConstructed;
     }
